+++ src/indicators/r058.rs
@@ -35,6 +35,7 @@ impl Calculate for R058 {
             for (ocid, ratio) in &item.second_lowest_bid_ratios {
                 if *ratio >= upper_fence {
                     set_result!(item, OCID, ocid, R058, *ratio);
+                    set_result!(item, Tenderer, item.winner_and_lowest_non_winner[ocid][0], R058, 0.0);
                 }
             }
         }